fix jdbc
